+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 - 🔒️(backend) add validation of Room.configuration
 - ✨(helm) add support multiple transcribe worker / endpoint #1247
 - ✨(backend) make LiveKit Egress recording encoding configurable #1288
+- ✨(summary) add speaker-to-participant assignment
 
 ### Changed